+++ R/plots.R
@@ -10,18 +10,27 @@
 #' @export
 #'
 #' @examples
-fpr_plot_fish_box <- function(dat = fish_abund, sit = my_site, theme = ggplot2::theme_bw()){
+# fpr_plot_fish_box <- function(dat = fish_abund, sit = my_site, theme = ggplot2::theme_bw()){
+fpr_plot_fish_box <- function(dat = fish_abund, sit = my_site, theme = ggplot2::theme_bw()) {
   dat |>
     dplyr::filter(
       site == sit &
         species_code != 'NFC'
     ) |>
     ggplot2::ggplot(ggplot2::aes(x = location, y = density_100m2)) +
-    ggplot2::geom_boxplot() +
+    ggplot2::geom_boxplot(color = "black", fill = NA) +
+    ggplot2::geom_dotplot(
+      binaxis = 'y',
+      stackdir = 'center',
+      dotsize = 1,
+      fill = "black",
+      color = "black"
+    ) +
     ggplot2::facet_grid(species_code ~ life_stage, scales = "fixed", as.table = TRUE) +
     theme +
-    ggplot2::theme(legend.position = "none", axis.title.x = ggplot2::element_blank()) +
-    ggplot2::geom_dotplot(binaxis = 'y', stackdir = 'center', dotsize = 1) +
+    ggplot2::theme(
+      legend.position = "none",
+      axis.title.x = ggplot2::element_blank()
+    ) +
     ggplot2::ylab(expression(Density ~ (fish / 100 ~ m^2)))
 }
-